// GGK-FoodTracker/src/app/Models/previous-orders.ts
export class PreviousOrder {
    orderID: number;
    itemName: string;
    status: string;
    orderDate: string;
    employeeID: number;
  }

  // [{"orderID":4,"itemName":"Chapathi","status":"Pending","orderDate":"2019-02-26","employeeID":5},{"orderID":5,"itemName":"Meals","status":"Pending","orderDate":"2019-02-26","employeeID":5},{"orderID":6,"itemName":"Meals","status":"Pending","orderDate":"2019-02-26","employeeID":5},{"orderID":7,"itemName":"Meals","status":"Pending","orderDate":"2012-06-18","employeeID":5},{"orderID":8,"itemName":"Meals","status":"Pending","orderDate":"2019-06-18","employeeID":5}]